Data binding and observability
A data binding connects data from a custom element (the host element)
to a property or attribute in its local DOM. The host element data can
be a property or sub-property represented by a data path, or data
generated based on one or more paths. A data path has the following
presentation:

    {{ hostProperty.subProperty }}

Let's discuss the observability of data changes.

An observable change is a data change that Polymer can associate with
a path. Certain changes are automatically observable:

  1. Setting a direct property of the element

         this.owner = 'Jane';

  2. Setting a subproperty of the element using a two-way data binding.

         <local-dom-child name="{{ hostProperty.subProperty }}">
         </local-dom-child>

     If `<local-dom-child>` makes a change to its `name` property, the
     change propagates up to the host, as a change to the path
     `hostProperty.subProperty`.

An unobservable change is a data change that Polymer cannot associate
with a path: changes that imperatively mutate an object or array.

  1. Setting a subproperty of an object:

         // unobservable subproperty changes
         this.reservation.from = 'Paris';
         this.reservation.to = 'Guangzhou';

     Changing the subproperty `reservation.from` here does not invoke
     the setter on `reservation`, so it is not automatically observable.

  2. Mutating an array:

         // unobservable change using native Array.push
         this.users.push({ name: 'Tom' });

     Mutating the arry does not invode the setter on `users`, so it is
     not automatically observable.

In both cases, you need to use Polymer methods to ensure that the
changes are observable.