Custom change notification events
Native elements like <input> don't provide the change notification
events that Polymer uses for upward data flow. To support two-way data
binding of native input elements, Polymer lets you associate a
**custom change notification event** with a data binding. For example,
when binding to a text input, you could specify the `input` (1) or
`change` (2) event:

    <input value="{{ firstName::change }}">

In this example, the `firstName` property is bound to the input's
`value` property. Whenever the input element fires its `change` event,
Polymer updates the `firstName` property to match the input value, and
invokes any associated property effects.

When querying bound data, do not forget the keyword `this`, which is
mandatory to obtain the property of the current object; Use one-way
binding for values in content, because they are only for disply purpose.

1. https://developer.mozilla.org/en-US/docs/Web/Events/input
2. https://developer.mozilla.org/en-US/docs/Web/Events/change